Add initial C-Suite project files
This change was previously committed in ag/9433397, but was reverted since it
breaks several build targets in the master-without-vendor branch. The breakage
was due to having a dependency on google-tradefed, which is only available in a
vendor project, and so not available in master-without-vendor. This dependency
is no longer included.

See go/csuite-proposal.

Test: make csuite, csuite-tradefed, run csuite-unit-tests
Change-Id: If81fa8985d575c2726ed8e14e4acf3936b1f465e