Add filtering to C-Suite test runner
C-Suite contains a launch test runner, which currently
runs all the app launch compatibility tests for all the
available apps. This change adds test filtering in the
launch runner by implementing ITestFilterReceiver to
help select a given set of tests on the command line,
test plans and ATP test configurations. This makes it
possible to use the test suite for only running the
test for relevant apps and bisecting an individual app
failure.

Test: make csuite, csuite-tradefed, run csuite-harness-tests
Change-Id: I318516f125f55c5ad8f38e557531aca0b6ebce8d