Add an app launch instrumentation test
Create a new instrumentation test runner in C-Suite
that verifies that an app can be successfully launched.
This differs from the existing runner in that it only
runs for a single app and not for apps defined in a
directory.

This change will eventually enable C-Suite to define a
single module per app.

Test: make csuite, csuite-tradefed, run csuite-harness-tests
Change-Id: Ibc55fa2e42cd14825414707e171fbf9b0c785a33