Configure C-Suite to only run on one suite ABI
Configure the suite to only run tests for the device's primary ABI since the framework runs on Java and does not contain architecture-dependent code. If not set, the suite's tests are needlessly rerun on all available ABIs.

Note that this change does not impact which ABI tested apps use.

Test: make csuite, csuite run launch
Change-Id: Idddf8194c3c1b116eac566516b110c8f5feca107
Fix: 144936076