Replace sharedUserId with adoptShellPermissionIdentity
Use adoptShellPermissionIdentity to launch the app instead of using
sharedUserId which requires signing the instrumentation with the
platform certificate.

The major reason for this change is that the suite could be built and
signed with a certificate that is different from the one used on the
device they're run on. This, in turn, leads to errors due to
incompatible key signaturess.

The secondary reason for this change is that sharedUserId was
deprecated in API level 29.

Test: make csuite, csuite-tradefed run launch, also verified on cutf
      and taimen in test lab using forrest
Change-Id: I77d9f470efc47a3ba2f7b28834ccf7a7b1121f5f
Fix: 145213933