Create a preparer to download and install an app
Apps were previously downloaded and installed in the
test runner which didn't clean up when it crashed. This
change introduces a new preparer that downloads and
installs the apps to properly free up resources.

Using a preparer also allows reusing downloaded APKs
between the different tests in a module and sharing the
implementation code of standard preparers.

Test: make csuite, csuite-tradefed run launch,
run csuite-harness-test
Change-Id: I72fd4bc6f4b0401451f8e4144d3a28423e554fee