Support passing options to AppSetupPreparer in test modules
Create a preparer that receives and stores test plan or command-line
config options in the IBuildInfo object. Currently it only supports
receiving the OPTION_GCS_APK_DIR of AppSetupPreparer. This is required
since TradeFed does not yet support passing options to preparers inside
test modules, and this preparer is specified on test plan level.

Test: Added unit tests
Change-Id: I5b0da9525e73aa36465a578c78b5f065a7d6a165