Configure project to run tests using atest
Simplify running tests which prior to this change had to be run from the
Android source root dir and required typing in several commands. This
was unecessary as the tests are plain-old host-side JUnit tests that can
be run without the C-Suite TradeFed runner.

The plans were converted to top-level module test configs deprecating
the old method of running the tests via the runner. Note that, besides
being unecessarily versbose, certain configuration elements such as the
reporters were removed as they cause issues with the atest runner.

This change also adds the tests to the postsubmit test group which is a
prerequisite for eventual inclusion in presubmit. Postsubmit tests are
not automatically run and require passing the :postsubmit option.

Change-Id: Idcc121c299e65d32529c31bf17fd839b8ce9305e
Test: atest --include-subdir :postsubmit :-)