Enable running Python unit tests using atest
Add Python unit tests to the project's build to simplify running them
both interactively and in headless mode. Certain minor, self-describing
code changes were required to integrate with the build/test system.

Worth noting is that the xml validity check for generated test modules
no longer validates against a schema since it requires an external
dependency which is not available. Importing this dependency is not
worth the hassle and maintenance burden since the schema we were using
was defined in the test itself and not really the official TradeFed
version.

Test: atest :postsubmit
Change-Id: I3b42de1526b2e446bdcdc4ec3e08cb700283e806