Add integration tests
Add integration tests that run the harness on devices to verify
end-to-end behavior that is not unit testable. This also provides us
with some level of protection against underlying dependency changes
that have negatively impacted us in the past.

The tests can be run using atest as indicated in this change's commit
Test stanza.

Test: atest --include-subdirs :postsubmit
Bug: 147836159
Change-Id: I06b733ae2ef866816feaadf4ce2b00f4d1e7edb9