Revert "Revert "Add exponential backoff to retries"" with fix
This reverts commit 211737853b11d1b9598dabd459dd74669175082f.

Reason for revert: The original CL ag/10842546 was reverted because it was missing a @Test annotation which is required in strict compilation mode. The CL was tested locally but was not verified by presubmit because there was another boot error breakage at the time. This revert of the revert comes with a fix on the original CL.

Test: presubmit
Change-Id: Ia7b48e55965c07d1096af06ee6ee1a883e3b3e36