Add features to help debugging test failures
Add options to preserve the test harness directories and log the test
command.

Note that this change aligns the behavior of methods that run commands
to raise an exception for processes that exit with a non-zero code. This
should have no impact on current tests since they all expect a
successful exit unless specified.

Test: atest --include-subdirs :all
Bug: 156700051
Change-Id: Ie08fe1f7f378d0a75e88667c605039ab828dc7ec