Refactor tests to improve readability
Replace shared fixture setup that uses instances variables with inline
setup. This helps eliminate most occurrences of the 'Mystery Guest' test
smell where we are not able to see the cause and effect between fixture
and verification logic because part of it is done outside the test
method. This also makes it harder to use the wrong fixture object which
is error-prone.

Test: atest --include-subdirs :all
Bug: 148404241
Change-Id: I38b8fa2c5b6adfcbdcf7576070c5e4a295a6f903