Update the system app removal logic
Previously, the system app remover removes an system app with the framework turned on and therefore has to modify the packages.xml manifest file in the system. This method is not robust enough to adapt to the manifest file format changes. This CL updates the uninstallation flow to remove the package installation with framework turned off and additionally uninstalls package updates in the beginning. Modifying packages.xml is no longer required.

Test: atest :presubmit
Test: forrest result https://android-build.googleplex.com/builds/forrest/run/L62500000796308187
Fix: 176043526
Change-Id: Ife9abc9828ee88c24331addcd3fee74a09d9cf0a