Hide instrumentation app from templates
The C-Suite instrumentation app is an internal tool in C-Suite and should be hidden from the user's view if possible. Previously each template specifies the instrumentation app installation config on the module level. This not only is confusing to the users but is also redundant since each test is installing the same apk repeatedly. This change moves the config from each template to the plan base config so that it becames transparent to the users. It also avoided duplicated installations.

Test: presubmit
Change-Id: Ib70d65118c5dc36e75dc433962a35bc7c287c41a