Check whether a system app is installed before trying to uninstall it
Previously the uninstaller attempts to uninstall a given app as a first step. While the uninstaller is designed to handle uninstalling non-existent apps, the log might cause confusion as the adb command would throw an error. It also triggers unnecessary error handling in the device's package manager. This CL adds a check and avoid attempting to uninstall a package that doesn't exist.

Test: atest :presubmit
Change-Id: Ie10c140d7ec4ef059b1d8a2c699600b4675d3698