Split app launching and crash checking into two instrumentation tests
Split the launching and crash checking logic apart from the original instrumentation tests into two instrumentation tests. This is an intermediate step along the effort of switching the test from device side to host driven. This CL is focused on refactoring and is created with minimal changes to the test logic. The classes affected in this CL will be further refactored/optimized in later changes.
The changes to test logic in this CL include:
1) The dropbox check device start time is now acquired from the host side script and passed into the crash checker. This will not affect the test results.
2) Previously when the launch intent for an app is not found, the test will not check for crash and report test passed. This CL makes the crash check happen anyway and will mark the test as failed in such cases. This is an intended change to the test logic.

Test: make csuite, atest csuite-harness-tests, csuite-tradefed run app-launch-test with a passing and a failing apps and the results are the same before and after this change.
Bug: 190093453
Change-Id: Ib72af0982baec3d7b1c031f36b192ec84d65c87c