Add package names to the prefix of crawler output file names
In a TradFed inevocation output page, all module outputs are listed together. If we don't have package names as the file name prefixes then the output files are mixed together. Adding the prefix makes it clear which zip file belongs to which package. In order to know the package name, we need to pass in the package name to the collect output method, which looks unnecessary as we can just save the package name in the class fields. Hence this CL refactored the class constructor to store the package name in a final field.

* Create the crawl tester with parameters instead of passing parameters in the startCrawl method.
* Use the apckage name as file name prefix for crawler output files.

Test: atest csuite-harness-tests
Bug: 192284237
Change-Id: Icbb0beaa89b5e111f9d8d1d764f06b3567f974d2
(cherry picked from commit 9c31f9f08084e3934394cefbb92c748203c5c1a8)
Merged-In: Icbb0beaa89b5e111f9d8d1d764f06b3567f974d2