Checks dropbox crash log even if the crawler command failed
Previously, when the crawler command failed, the crawl test exit before checking dropbox for app crash data, making it hard to tell whether the test crash is a cause of the crawler error or app crash. This change makes the crawl test check both cases independently and combine the results.

Test: make csuite
Bug: 206528595
Change-Id: I17070e798141dc6abb481d6aff51dc7fae268769