Fix a default value a integration test util method
The default value for the shell command check argument should be false because: 1) The test code already asserts on the return value 2) If error is raised from the subprocess check then there would not be enough error details for debugging. 3) To be consistant with the default value in python subprocess module.

Test: presubmit
Bug: 210566308
Change-Id: I60d1d6264d241cd57ef96164fca12bc9c11cf8af