Rename the crawl test preparer to host preparer
The app crawl test preparer is a host preparer as it only prepares the host for crawling apps. Since it is executed on the suite level, we cannot assume any actions made to the device in the future will remain their effects between retires. Therefore renaming the preparer to host preparer makes it clearer that only host actions can be added in this class.

Test: presubmit
Change-Id: I33ff65dc0664266d59e28da7d42aedae7bb538e8