Support putting template files in subdirectories
Since we now accept extra templates, we need to support the case when template files are put into subdirectires and referred by their relative path in template mappings.
For example, a developer could put a template file in template_group1/template1.xml and use it in the template mapping: module1 -> mytemplate_group1/template1.
If we do not support it, then we would need to throw an error when template files are put in directories.

Test: atest csuite-harness-tests
Bug: 197867682
Change-Id: I2dfda902fbc362f1c74e42b2e1cd09d9d02e8703