Add a push file preparer to the suite base config
Add a push file preparer to the suite base config so that tests can be configured to push files to the device before tests without having to create a new test suite plan. Note that this addition has no effect on existing tests if no parameters are supplied to the new preparer.

Test: presubmit
Fix: 249184504
Change-Id: I4fd54ea27503fbb74b5c58c03085004ebb36c77c