Uninstall the test package before installing it
Uninstall the test package before installing it in case it was not properly cleaned between retries. This CL also removes the command to reset the package as it is no longer necessary. This also fixes a bug, as previously we were reseting the package after installation which causes granted permissions getting revoked. The permission can be fixed by move the reset command before the install command but uninstalling is better than reset as uninstalling can unblock the installation if a higher version of the apk has already been installed.

Test: presubmit
Bug: 238972145
Change-Id: Id1fd2d4c6ede9352f14c22d2c8cc92d86a1c9364