Parse an apk's package name and uninstall it before installing the package
Currently the apk installer attempt to install an apk without attempting to uninstall it first. If the package was already installed on the device with a newer version or different signature, the installation will fail.
This CL adds an attempt to uninstall the package before the installation.

This CL also removes the uninstalling of the subject package at the beginning of the launch test since the package will be uninstalled by the apk installer.

Fix: 256993764
Test: https://android-build.googleplex.com/builds/abtd/run/L21000000957313023
Change-Id: I12ca84b01c0831680d140954ebc070056c17ef25