Use pm and am commands to launch an app
Previously we use the monkey command to launch an app, but the monkey process can sometimes be flaky, causing false positives in tests.
This CL changed the device util to use the pm and am commands to start an app.

Additionally, previously when we use monkey to launch an app we were only looking for any activity that has the LAUNCHER category, but this is not accurate as some apps may have multiple launcher activities not none. This CL adds a logic to sort the activities according to their likelihood to be the main entry point of the app.

Test: unit test
Fix: 267823695
Fix: 253502086
Change-Id: Iaa75f1bb3f79ceae5aca187e8d4c1779e14c133e