Launch apps using 'monkey' command as primary method
In the past, we stopped using the 'monkey' command due to its occasional flakiness. However, we have discovered that the 'pm/am' method for launching apps may not work correctly when the app has encrypted or obfuscated activity names. To address this issue, this CL changes the default method for launching apps to use the 'monkey' command as the primary method. If the 'monkey' command fails to launch the app, the tests will then fallback to the 'pm/am' method.

This update should improve the reliability of our app tests, particularly for apps with encrypted or obfuscated activity names.

Test: unit tests
Fix: 269345885
Change-Id: I6ef5dfae44f011e3b36d5c86910bf9bf4b55d5a4