Use SIGINT to stop `screenrecord` process and wait for process termination
Modifies the termination of the `screenrecord` process by replacing Process.destroy() with a SIGINT signal and waiting for the process to terminate. Previously, Process.destroy() was used, which sent a SIGTERM signal and did not wait for the process to terminate, resulting in incomplete MP4 files being produced occasionally. This change sends a SIGINT signal to the process to mimic the behavior of pressing Ctrl+C when running the `screenrecord` command from the command line, and also waits for the process to terminate to avoid incomplete MP4 files.

Fix: 247054178
Test: forrest
Change-Id: I84c0924041396da21cddb24bb1693fe43051e945