Extract all the crawler options to a separate class
Extract all the options out from the test script so that new crawl test scripts won't have to re-define all the crawler options. Most of the options are internal to the crawler and is not specific to the test script. After this change, the test script can still override the options if needed.
The deprecated options can be removed once all the scheduled prod configurations are updated.

Test: local unit test and crawl test
Change-Id: If3746bb80f684cf317f0c166d559c43daab5440b