Webview test use a new instance of crawler for each run
Currently each instance of the app crawl tester only support at most 1 run of the crawler. The webview test currently contains an verification step that reruns the crawler with a different webview version. The verification run step always fail because the crawler would throw an exception, and the test will pass because the verification run fails, making the test ineffective. This CL fixes this issue.

Test: presubmit
Change-Id: Ib048414136737648f9e49b24a782b514b4b23f3d