Using "cmd package resolve-activity" to get the launch activity for a package.
This is a more reliable way to get the launch activity for a package than the "pm dump" command, which can be flaky. For example for Youtube app current getLaunchActivityName method does not provide the right launch activity for the app.

Added unit tests for the new method.

Test: ran unit tests locally.
Flag: TEST_ONLY
Bug: 410485139
Change-Id: I0feb980d96da9e92ed757160b4a49b77e521a640